+++ src/components/UI/AppInput.vue
@@ -21,7 +21,5 @@ export default {
   width: 100%;
   border: 2px solid rgb(187, 84, 32);
   padding: 10px 15px;
-  margin-right: 15px;
-  margin-top: 15px;
 }
 </style>
