+++ src/components/UI/index.js
@@ -1,7 +1,7 @@
 import MyButton from '@/components/UI/AppPrimaryButton'
-import MyInput from '@/components/UI/MyInput'
-import MyDialog from '@/components/UI/MyDialog'
-import MySelect from '@/components/UI/MySelect'
+import MyInput from '@/components/UI/AppInput'
+import MyDialog from '@/components/UI/AppDialog'
+import MySelect from '@/components/UI/AppSelect'
 
 export default [
   MyButton,
